Criação da classe RequestService, que controla as requisições a web e erros relacionados.
