+++ lib/data/repositories/covid19_api.dart
@@ -3,20 +3,26 @@ import 'package:covid19_tracker_in_flutter/data/models/continent_summary_model.d
 import 'package:covid19_tracker_in_flutter/data/models/country_summary_model.dart';
 import 'package:covid19_tracker_in_flutter/data/models/historical_model.dart';
 import 'package:covid19_tracker_in_flutter/data/models/summary_model.dart';
+import 'package:covid19_tracker_in_flutter/data/services/request_service.dart';
 import 'package:covid19_tracker_in_flutter/domain/contracts/covid19_api_contract.dart';
-import 'package:dio/dio.dart';
 
 const BASE_URL = 'https://disease.sh/v3/covid-19';
 
 class Covid19Api implements Covid19ApiContract {
+  RequestService _service;
+
+  Covid19Api(RequestService requestService) {
+    _service = requestService;
+  }
+
   Future<SummaryModel> worldSummary() async {
     final String request = BASE_URL + '/all';
-    return SummaryModel.fromMap(await _doRequest(request));
+    return SummaryModel.fromMap(await _service.request(request));
   }
 
   Future<List<ContinentSummaryModel>> continentsSummary() async {
     final String request = BASE_URL + '/continents';
-    final List response = await _doRequest(request);
+    final List response = await _service.request(request);
     return List.generate(
       response.length,
       (index) => ContinentSummaryModel.fromMap(response[index]),
@@ -26,7 +32,7 @@ class Covid19Api implements Covid19ApiContract {
   Future<ContinentSummaryModel> findContinentSummary(String continent) async {
     _validateStringParameter(continent);
     final String request = '$BASE_URL/continents/$continent';
-    return ContinentSummaryModel.fromMap(await _doRequest(request));
+    return ContinentSummaryModel.fromMap(await _service.request(request));
   }
 
   _validateStringParameter(String parameter) {
@@ -36,7 +42,7 @@ class Covid19Api implements Covid19ApiContract {
 
   Future<List<CountrySummaryModel>> countriesSummary() async {
     final String request = BASE_URL + '/countries';
-    final List response = await _doRequest(request);
+    final List response = await _service.request(request);
     return List.generate(
       response.length,
       (index) => CountrySummaryModel.fromMap(response[index]),
@@ -50,7 +56,7 @@ class Covid19Api implements Covid19ApiContract {
     _validateStringParameter(country);
     _validatePeriodParameter(period);
     final String request = '$BASE_URL/countries/$country${getPeriod(period)}';
-    return CountrySummaryModel.fromMap(await _doRequest(request));
+    return CountrySummaryModel.fromMap(await _service.request(request));
   }
 
   _validatePeriodParameter(Period period) {
@@ -67,7 +73,7 @@ class Covid19Api implements Covid19ApiContract {
   Future<HistoricalModel> worldHistorical(String numberOfDays) async {
     _validateNumberOfDays(numberOfDays);
     final String request = '$BASE_URL/historical/all?lastdays=$numberOfDays';
-    return HistoricalModel.fromMap(await _doRequest(request));
+    return HistoricalModel.fromMap(await _service.request(request));
   }
 
   void _validateNumberOfDays(String days) {
@@ -83,28 +89,6 @@ class Covid19Api implements Covid19ApiContract {
     _validateStringParameter(country);
     _validateNumberOfDays(numberOfDays);
     final request = '$BASE_URL/historical/$country?lastdays=$numberOfDays';
-    return HistoricalModel.fromMap(await _doRequest(request));
-  }
-
-  Future<dynamic> _doRequest(String request) async {
-    try {
-      final Response response = await Dio().get(request);
-      return _validateResponse(response);
-    } catch (error) {
-      _onError(error);
-      return null;
-    }
-  }
-
-  dynamic _validateResponse(Response response) {
-    if (response == null || response.statusCode != 200)
-      throw Covid19ApiError(apiErrorType: ApiErrorType.INVALID_RESPONSE);
-    return response.data;
-  }
-
-  void _onError(error) {
-    if (error.runtimeType == DioError)
-      throw Covid19ApiError(dioErrorType: error.type);
-    throw Covid19ApiError();
+    return HistoricalModel.fromMap(await _service.request(request));
   }
 }
